akatsuki/requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
mariadb==1.0.4
mysqlclient @ file:///C:/Users/Norman%20Rivas%20Fuentes/Desktop/Proyecto%20iwg/AkatsukiTracker/AkatsukiTracker/mysqlclient-1.4.6-cp38-cp38-win32.whl
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
